Add files via upload
This is a Banking data set which contains the information of customers who were targeted by the bank to provide personal loan. It is a classification problem where in which based on the features given, we classify whether the targeted customer will subscribe for the personal loan or not. It is a rich data set with 40,000 records and 40 features, out of which 20 features are categorical and 20 features are continuous. It contains very minimal missing values.